requirements.txt
# Speech-to-text
openai-whisper

# Text-to-speech
gTTS
playsound
pyttsx3 # for multiple voices 

# Audio recording & processing
sounddevice
soundfile
numpy
scipy

# LLM backend
ollama
